+++ test/unit/attachments.test.ts
@@ -188,50 +188,6 @@ config.parallel('attachments.test.ts', () => {
             });
             c.database.destroy();
         });
-        it('should not update the document if skipIfSame=true and same data', async () => {
-            const c = await humansCollection.createAttachments(1);
-            const doc = await c.findOne().exec(true);
-            const data = blobBufferUtil.createBlobBuffer(AsyncTestUtil.randomString(100), 'text/plain');
-            await doc.putAttachment({
-                id: 'cat.txt',
-                data,
-                type: 'text/plain'
-            }, true);
-            const revBefore = doc.revision;
-            await doc.putAttachment({
-                id: 'cat.txt',
-                data,
-                type: 'text/plain'
-            }, true);
-            await wait(50);
-            assert.strictEqual(
-                revBefore,
-                doc.revision
-            );
-
-            c.database.destroy();
-        });
-        it('should update the document if skipIfSame=true and different data', async () => {
-            const c = await humansCollection.createAttachments(1);
-            const doc = await c.findOne().exec(true);
-            await doc.putAttachment({
-                id: 'cat.txt',
-                data: blobBufferUtil.createBlobBuffer(AsyncTestUtil.randomString(100), 'text/plain'),
-                type: 'text/plain'
-            }, true);
-            const revBefore = doc.revision;
-            await doc.putAttachment({
-                id: 'cat.txt',
-                data: blobBufferUtil.createBlobBuffer(AsyncTestUtil.randomString(100), 'text/plain'),
-                type: 'text/plain'
-            }, false);
-            await wait(50);
-            assert.notStrictEqual(
-                revBefore,
-                doc.revision
-            );
-            c.database.destroy();
-        });
     });
     describe('.getAttachment()', () => {
         it('should get the attachment', async () => {