+++ test/unit/rx-storage-implementations.test.ts
@@ -24,7 +24,8 @@ import {
     ById,
     stackCheckpoints,
     defaultHashFunction,
-    deepFreeze
+    deepFreeze,
+    stripAttachmentsDataFromDocument
 } from '../../';
 import Ajv from 'ajv';
 import {
@@ -58,7 +59,6 @@ import {
 import { compressObject } from 'jsonschema-key-compression';
 
 import {
-    hashAttachmentData,
     getAttachmentSize
 } from '../../plugins/attachments';
 
@@ -762,7 +762,7 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 });
 
                 const key = 'foobar';
-                let docData: RxDocumentWriteData<TestDocType> = {
+                let docData: RxDocumentData<TestDocType> = {
                     key,
                     value: 'barfoo1',
                     _attachments: {},
@@ -784,7 +784,7 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 assert.deepStrictEqual(res1.error, {});
 
                 // change once
-                let newDocData: RxDocumentWriteData<TestDocType> = clone(docData);
+                let newDocData: RxDocumentData<TestDocType> = clone(docData);
                 newDocData._meta.foobar = 1;
                 newDocData._meta.lwt = now();
                 newDocData._rev = createRevision(defaultHashFunction, newDocData, docData);
@@ -884,7 +884,7 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 });
 
                 // insert
-                const docData: RxDocumentWriteData<TestDocType> = {
+                const docData: RxDocumentData<TestDocType> = {
                     key: 'foobar',
                     value: 'barfoo1',
                     _deleted: false,
@@ -2328,7 +2328,6 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                     'text/plain'
                 );
                 const dataStringBase64 = await blobBufferUtil.toBase64String(dataBlobBuffer);
-                const attachmentHash = await hashAttachmentData(dataStringBase64);
                 const dataLength = getAttachmentSize(dataStringBase64);
 
                 const writeData: RxDocumentWriteData<TestDocType> = {
@@ -2341,7 +2340,6 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                     },
                     _attachments: {
                         foo: {
-                            digest: 'md5-' + attachmentHash,
                             length: dataLength,
                             data: dataStringBase64,
                             type: 'text/plain'
@@ -2387,7 +2385,6 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 );
 
                 const dataStringBase64 = await blobBufferUtil.toBase64String(dataBlobBuffer);
-                const attachmentHash = await hashAttachmentData(dataStringBase64);
                 const dataLength = getAttachmentSize(dataStringBase64);
 
                 const writeData: RxDocumentWriteData<TestDocType> = {
@@ -2400,7 +2397,6 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                     },
                     _attachments: {
                         foo: {
-                            digest: 'md5-' + attachmentHash,
                             length: dataLength,
                             data: dataStringBase64,
                             type: 'text/plain'
@@ -2419,7 +2415,6 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 await waitUntil(() => flattenEvents(emitted).length === 1);
 
                 assert.strictEqual(writeResult._attachments.foo.type, 'text/plain');
-                assert.strictEqual(writeResult._attachments.foo.digest, 'md5-' + attachmentHash);
 
                 /**
                  * When getting the document from the storage again,
@@ -2495,9 +2490,7 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 let previous: RxDocumentData<TestDocType> | undefined;
 
                 const dataBlobBuffer = blobBufferUtil.createBlobBuffer(randomString(20), 'text/plain');
-                const dataString = await blobBufferUtil.toBase64String(dataBlobBuffer);
                 const dataStringBase64 = await blobBufferUtil.toBase64String(dataBlobBuffer);
-                const attachmentHash = await hashAttachmentData(dataStringBase64);
                 const writeData: RxDocumentWriteData<TestDocType> = {
                     key: 'foobar',
                     value: 'one',
@@ -2508,9 +2501,8 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                     },
                     _attachments: {
                         foo: {
-                            digest: 'md5-' + attachmentHash,
                             length: blobBufferUtil.size(dataBlobBuffer),
-                            data: dataString,
+                            data: dataStringBase64,
                             type: 'text/plain'
                         }
                     }
@@ -2536,12 +2528,9 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 writeData._attachments = flatClone(previous._attachments) as any;
 
                 const data2 = blobBufferUtil.createBlobBuffer(randomString(20), 'text/plain');
-                const dataStringBase642 = await blobBufferUtil.toBase64String(data2);
-                const attachmentHash2 = await hashAttachmentData(dataStringBase642);
                 const dataString2 = await blobBufferUtil.toBase64String(data2);
                 writeData._attachments.bar = {
                     data: dataString2,
-                    digest: 'md5-' + attachmentHash2,
                     length: blobBufferUtil.size(data2),
                     type: 'text/plain'
                 };
@@ -2582,8 +2571,6 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 });
 
                 const data = blobBufferUtil.createBlobBuffer(randomString(20), 'text/plain');
-                const dataStringBase64 = await blobBufferUtil.toBase64String(data);
-                const attachmentHash = await hashAttachmentData(dataStringBase64);
                 const dataString = await blobBufferUtil.toBase64String(data);
                 const writeData: RxDocumentWriteData<TestDocType> = {
                     key: 'foobar',
@@ -2595,7 +2582,6 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                     },
                     _attachments: {
                         foo: {
-                            digest: 'md5-' + attachmentHash,
                             length: blobBufferUtil.size(data),
                             data: dataString,
                             type: 'text/plain'
@@ -2612,7 +2598,10 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 deleteData._rev = EXAMPLE_REVISION_2;
 
                 await storageInstance.bulkWrite(
-                    [{ previous: writeData, document: deleteData }],
+                    [{
+                        previous: stripAttachmentsDataFromDocument(writeData),
+                        document: deleteData
+                    }],
                     testContext
                 );